Skeleton of a possible presentation for Friday
Use it if you like it, if not, then don't.

Mock-up Interface has no image since I think it should be the mock-up itself.
Similar thing with the SAT toy/mock example, depending on our results it should look one way or another

The slides have a script (to read), modify it at your will.